Check for failures when creating or updating BCD stores under macOS.

+++ Darwin/bcd-sys.sh
@@ -121,19 +121,40 @@ $resdir/winload.sh "$1" "$2" "$3" "$4" "$5" "$6" "$7" "$8" "$9" "${10}" "${11}" 
 cp $resdir/Templates/BCD-NEW $tmpdir/BCD-Windows
 hivexregedit --merge --prefix BCD00000001 $tmpdir/BCD-Windows $resdir/Templates/winload.reg
 hivexsh -w -f $tmpdir/winload.txt $tmpdir/BCD-Windows
+if [[ ! $? -eq 0 ]]; then
+   echo -e "${RED}Failed to create main BCD store.${NC}"
+   return 1
+fi
 if   [[ "$3" == "uefi" ]]; then
      if [[ "$verbose" == "true" ]]; then echo "Build the recovery BCD store..."; fi
      $resdir/recovery.sh "$2" "$5" "$8" > $tmpdir/recovery.txt
      cp $resdir/Templates/BCD-NEW $tmpdir/BCD-Recovery
      hivexregedit --merge --prefix BCD00000001 $tmpdir/BCD-Recovery $resdir/Templates/recovery.reg
      hivexsh -w -f $tmpdir/recovery.txt $tmpdir/BCD-Recovery
+     if [[ ! $? -eq 0 ]]; then
+        echo -e "${RED}Failed to create recovery BCD store.${NC}"
+        return 1
+     fi
      if [[ "$verbose" == "true" ]]; then echo "Copy the BCD hives to the ESP folders..."; fi
      cp $tmpdir/BCD-Windows "$2"/EFI/Microsoft/Boot/BCD
+     if [[ ! $? -eq 0 ]]; then
+        echo -e "${RED}Failed to copy main hive to the ESP.${NC}"
+        return 1
+     fi
      cp $tmpdir/BCD-Recovery "$2"/EFI/Microsoft/Recovery/BCD
+     if [[ ! $? -eq 0 ]]; then
+        echo -e "${RED}Failed to copy recovery hive to the ESP.${NC}"
+        return 1
+     fi
 elif [[ "$3" == "bios" ]]; then
      if [[ "$verbose" == "true" ]]; then echo "Copy the main BCD hive to the boot folder..."; fi
      cp $tmpdir/BCD-Windows "$2"/Boot/BCD
+     if [[ ! $? -eq 0 ]]; then
+        echo -e "${RED}Failed to copy main hive to the system partition.${NC}"
+        return 1
+     fi
 fi
+return 0
 }
 
 # Build new Windows entry and update existing BCD stores.
@@ -143,12 +164,25 @@ if [[ "$verbose" == "true" ]]; then echo "Update main BCD hive with new entries.
 $resdir/winload.sh "$1" "$2" "$3" "$4" "$5" "$6" "$7" "$8" "$9" "${10}" "${11}" "${12}" > $tmpdir/winload.txt
 if   [[ "$3" == "uefi" ]]; then
      hivexsh -w -f $tmpdir/winload.txt "$2/EFI/Microsoft/Boot/BCD"
+     if [[ ! $? -eq 0 ]]; then
+        echo -e "${RED}Failed to update main BCD store.${NC}"
+        return 1
+     fi
      if [[ "$verbose" == "true" ]]; then echo "Update recovery BCD hive with new entries..."; fi
      $resdir/recovery.sh "$2" "$5" "$8" > $tmpdir/recovery.txt
-     hivexsh -w -f $tmpdir/recovery.txt "$2/EFI/Microsoft/Recovery/BCD" 
+     hivexsh -w -f $tmpdir/recovery.txt "$2/EFI/Microsoft/Recovery/BCD"
+     if [[ ! $? -eq 0 ]]; then
+        echo -e "${RED}Failed to update recovery BCD store.${NC}"
+        return 1
+     fi
 elif [[ "$3" == "bios" ]]; then
      hivexsh -w -f $tmpdir/winload.txt "$2/Boot/BCD"
+     if [[ ! $? -eq 0 ]]; then
+        echo -e "${RED}Failed to update main BCD store.${NC}"
+        return 1
+     fi
 fi
+return 0
 }
 
 # Copy the WBM files for the specified firmware and set the file attributes the same as bcdboot.
@@ -945,6 +979,10 @@ if  [[ "$firmware" != "uefi" && "$firmware" != "bios" && "$firmware" != "both" ]
     if  [[ "$virtual" == "true" ]]; then umount_vpart; fi
     exit 1
 else
+    if [[ -f $tmpdir/winload.txt || -f $tmpdir/recovery.txt || -f $tmpdir/wbmfwvar.txt || \
+          -f $tmpdir/BCD-Windows || -f $tmpdir/BCD-Recovery ]]; then
+       cleanup
+    fi
     if [[ "$firmware" == "uefi" || "$firmware" == "both" ]]; then
        if  [[ -z "$syspath" ]]; then
            get_syspath "$firmware" "$windisk"
@@ -1005,6 +1043,11 @@ else
             fi
             build_stores "$winpath" "$syspath" "$fwmode" "$setfwmod" "$createbcd" "$prewbmdef" \
                          "$prodname" "$locale" "$verbose" "$virtual" "$vrtpath" "$imgstring"
+            if [[ ! $? -eq 0 ]]; then
+               if [[ "$virtual" == "true" ]]; then umount_vpart; fi
+               if [[ "$rmsysmnt" == "true" ]]; then umount_system; fi
+               exit 1
+            fi
             if [[ "$setfwmod" == "false" && "$efibootvars" == "true" ]]; then
                if [[ ! -z "$wbmoptnum" ]]; then
                   if [[ "$verbose" == "true" ]]; then echo "Remove the current Windows Boot Manager option..."; fi
@@ -1035,6 +1078,11 @@ else
             fi
             build_stores "$winpath" "$syspath" "$fwmode" "$setfwmod" "$createbcd" "$prewbmdef" \
                          "$prodname" "$locale" "$verbose" "$virtual" "$vrtpath" "$imgstring"
+            if [[ ! $? -eq 0 ]]; then
+               if [[ "$virtual" == "true" ]]; then umount_vpart; fi
+               if [[ "$rmsysmnt" == "true" ]]; then umount_system; fi
+               exit 1
+            fi
             if [[ "$setfwmod" == "false" && "$efibootvars" == "true" ]]; then
                if [[ -z "$wbmoptnum" ]]; then
                   if [[ "$verbose" == "true" ]]; then echo "Add the Windows Boot Manager to the firmware..."; fi
@@ -1066,6 +1114,11 @@ else
             fi
             update_winload "$winpath" "$syspath" "$fwmode" "$setfwmod" "$createbcd" "$prewbmdef" \
                            "$prodname" "$locale" "$verbose" "$virtual" "$vrtpath" "$imgstring"
+            if [[ ! $? -eq 0 ]]; then
+               if [[ "$virtual" == "true" ]]; then umount_vpart; fi
+               if [[ "$rmsysmnt" == "true" ]]; then umount_system; fi
+               exit 1
+            fi
             if [[ "$setfwmod" == "false" && "$efibootvars" == "true" && -z "$wbmoptnum" ]]; then
                if [[ "$verbose" == "true" ]]; then echo "Add the Windows Boot Manager to the firmware..."; fi
                sudo bootoption create -l "$wbmefipath" -d "Windows Boot Manager" -@ $resdir/Templates/wbmoptdata.bin > /dev/null
@@ -1158,6 +1211,11 @@ else
           fi
           build_stores "$winpath" "$syspath" "$fwmode" "$setfwmod" "$createbcd" "$prewbmdef" \
                        "$prodname" "$locale" "$verbose" "$virtual" "$vrtpath" "$imgstring"
+       if [[ ! $? -eq 0 ]]; then
+          if [[ "$virtual" == "true" ]]; then umount_vpart; fi
+          if [[ "$rmsysmnt" == "true" ]]; then umount_system; fi
+          exit 1
+       fi
        elif [[ "$sysbtmgr" == "true" && "$clean" == "true" ]]; then
             if [[ "$verbose" == "true" ]]; then echo "Remove current BCD store..."; fi
             rm -f "$syspath"/Boot/BCD
@@ -1187,6 +1245,11 @@ else
             fi
             build_stores "$winpath" "$syspath" "$fwmode" "$setfwmod" "$createbcd" "$prewbmdef" \
                          "$prodname" "$locale" "$verbose" "$virtual" "$vrtpath" "$imgstring"
+           if [[ ! $? -eq 0 ]]; then
+              if [[ "$virtual" == "true" ]]; then umount_vpart; fi
+              if [[ "$rmsysmnt" == "true" ]]; then umount_system; fi
+              exit 1
+           fi
        else
             createbcd="false"
             if   [[ "$sysuwfver" != "NULL" && "$sysvhdver" != "NULL" && "$localuwfver" != "NULL" && "$localvhdver" != "NULL" ]]; then
@@ -1223,10 +1286,16 @@ else
             fi
             update_winload "$winpath" "$syspath" "$fwmode" "$setfwmod" "$createbcd" "$prewbmdef" \
                            "$prodname" "$locale" "$verbose" "$virtual" "$vrtpath" "$imgstring"
+            if [[ ! $? -eq 0 ]]; then
+               if [[ "$virtual" == "true" ]]; then umount_vpart; fi
+               if [[ "$rmsysmnt" == "true" ]]; then umount_system; fi
+               exit 1
+            fi
        fi
        if [[ "$rmsysmnt" == "true" ]]; then umount_system; fi
        if [[ "$virtual" == "true" ]]; then umount_vpart; fi
        cleanup
        echo "Finished configuring BIOS boot files."
     fi
+exit 0
 fi
